requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
beautifulsoup4==4.12.2
requests-html==0.10.0
pyppeteer==1.0.2
stripe==10.8.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
loguru==0.7.2
openai==1.3.0